third_party/WebKit/LayoutTests/fast/pagination/paged-x-column-gap.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  html {
    height: 500px;
  }
  body {
    margin: 0;
    column-gap: 0;
    overflow: -webkit-paged-x;
  }
  .green {
    background-color: green;
    height: 100px;
    break-after: page;
  }
</style>
<div class="green">Page 1<br><br>There should be a white gap between the two green pages</div>
<div class="green">Page 2</div>
<script>
  test(() => {
    assert_equals(document.documentElement.scrollWidth, 2*document.documentElement.clientWidth,
        "Scrollable width should be twice the page width with no column-gap.");
  }, "Test for initial column-gap: 0.");

  test(() => {
    document.body.style.columnGap = "normal";
    assert_greater_than(document.documentElement.scrollWidth, 2*document.documentElement.clientWidth,
        "Scrollable width should be larger than twice the page width with column-gap.");
  }, "Changing to column-gap: normal should create a gap between the pages");
</script>
